refactor: use `AstSymbol` to store infomation related to `ReferenceId`
