+++ crates/rolldown/src/bundler/utils/mod.rs
@@ -1,4 +1,5 @@
 pub mod ast_scope;
+pub mod ast_symbol;
 pub mod bitset;
 pub mod renamer;
 pub mod resolve_id;
